Add dependency for the program
Visual C++ Redistributable for Visual Studio 2015 contains the Universal C
Runtime (ucrtbase.dll) which is required to run the binary if it's built from
Python 3.5. This should be installed automatically with a Windows update,
but if someone doesn't have it then they can manually install it with
vc_redist.x64.exe.