Finish the TCP socket test GUI, only server side
I imagine the computer w/ the GUI will be the server but it should be
fairly simple to change. This implementation of the host freezes the GUI
while waiting for the accept call to return, so making it asynchronous or
putting it on a separate thread should be done later. I'm investigating
the different methods of doing so and how to integrate it with Qt.

However, for now this is a functional way to test if the sockets are
communicating correctly.

Also added some more info to the README.