Further WIP asynchronous server work
GUI is now the receiver of messages asynchronously and doesn't hang while
waiting for a connection. This uses signals and slots aka callback
methods/functions. Trying to use asyncio with the Qt Event Loop proved to
be too experimental and demanding to be worth it when we aren't going to
have any insane callback chains.

However, it currently doesn't receive and accept messages. If you try to
send a message from the GUI to client it will crash. The socket
communication will need to be reworked...